psalminized

+++ src/Service/TemperatureProvider/Hdd.php
@@ -4,12 +4,15 @@ namespace Xuedi\PhpSysMon\Service\TemperatureProvider;
 
 use Xuedi\PhpSysMon\LinuxPath;
 
+/**
+ * @psalm-suppress ForbiddenCode
+ */
 class Hdd implements TemperatureProvider
 {
     public function getTemperature(LinuxPath $device): int
     {
         $command = "sudo hddtemp --unit=C --wake-up --numeric " . $device->asString();
 
-        return shell_exec($command);
+        return (int)shell_exec($command);
     }
 }
